training/Audio/keyword_detect/requirements.txt
absl-py==0.8.1
astor==0.8.1
audioread==2.1.8
awscli==1.16.300
botocore==1.13.36
cffi==1.13.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
easydict==1.9
gast==0.3.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
jmespath==0.9.4
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.6.3
llvmlite==0.30.0
Markdown==3.1.1
matplotlib==3.1.2
numba==0.46.0
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.3
protobuf==3.11.1
pyasn1==0.4.8
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
resampy==0.2.2
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
SoundFile==0.10.3.post1
sox==1.3.7
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.41.1
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
